+++ R/S3-methods.R
@@ -2008,7 +2008,7 @@ getME.brmsfit <- function(object, name, ...) {
   rv <- NULL
   if (name == "X") {
     rv <- as.matrix(cbind(1, model_frame(object)[pred_vars(object, fe.only = T)]))
-    colnames(rv)[1] = "(Intercept)"
+    colnames(rv)[1] = "Intercept"
   }
   rv
 }